k8s: extract secret syncer into pkg/envoy
This commit moves the Envoy secret syncer into the package
`pkg/envoy`.
