+++ pkg/envoy/secretsync_test.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Authors of Cilium
 
-package watchers
+package envoy
 
 import (
 	"testing"
